+++ src/error/mod.rs
@@ -7,6 +7,5 @@
 pub mod control_syn;
 /// 変数や関数などの定義に関するログ
 pub mod define_msg;
-pub mod err_handling;
 pub mod log;
 pub mod parse_err;