+++ src/parse/expr/mod.rs
@@ -10,3 +10,6 @@ use crate::error::parse_err;
 use crate::token::token;
 use crate::parse::node;
 use crate::parse::resp;
+use parse_comper::parse_comper_op;
+use parse_trim::parse_trim;
+use parse_factor::parse_factor;