+++ src/parse/expr/parse_expr.rs
@@ -1,17 +1,17 @@
-use crate::token::token;
-use crate::parse::node;
-use crate::parse::resp;
-use crate::parse::expr::parse_trim::parse_trim;
+use super::*;
 
 
-pub fn parse_expr(tokens: Vec<token::Token>, index: &mut usize) -> node::CalculNode {
-    let mut node = parse_trim(tokens.clone(), index);
+pub fn parse_expr(
+        tokens: Vec<token::Token>,
+        index: &mut usize
+) -> Result<node::CalculNode, parse_err::ParseErrs> {
+    let mut node = parse_trim(tokens.clone(), index)?;
 
     while tokens.len() > *index {
         match tokens[*index].kind {
             token::TokenKind::TokenAdd => {
                 *index += 1;
-                let right = parse_trim(tokens.clone(), index);
+                let right = parse_trim(tokens.clone(), index)?;
                 node = resp::handler::make_operator_node(
                     node,
                     right,
@@ -20,7 +20,7 @@ pub fn parse_expr(tokens: Vec<token::Token>, index: &mut usize) -> node::CalculN
             }
             token::TokenKind::TokenSub => {
                 *index += 1;
-                let right = parse_trim(tokens.clone(), index);
+                let right = parse_trim(tokens.clone(), index)?;
                 node = resp::handler::make_operator_node(
                     node,
                     right,
@@ -36,5 +36,5 @@ pub fn parse_expr(tokens: Vec<token::Token>, index: &mut usize) -> node::CalculN
         }
     }
 
-    return node;
+    return Ok(node);
 }