+++ src/parse/expr/parse_trim.rs
@@ -1,17 +1,17 @@
-use crate::parse::node;
-use crate::token::token;
-use crate::parse::expr::parse_factor::parse_factor;
-use crate::parse::resp;
+use super::*;
 
 
-pub fn parse_trim(tokens: Vec<token::Token>, index: &mut usize) -> node::CalculNode {
-    let mut node = parse_factor(tokens.clone(), index);
+pub fn parse_trim(
+        tokens: Vec<token::Token>,
+        index: &mut usize
+) -> Result<node::CalculNode, parse_err::ParseErrs> {
+    let mut node = parse_factor(tokens.clone(), index)?;
 
     while tokens.len() > *index {
         match tokens[*index].kind {
             token::TokenKind::TokenMul => {
                 *index += 1;
-                let right = parse_factor(tokens.clone(), index);
+                let right = parse_factor(tokens.clone(), index)?;
                 node = resp::handler::make_operator_node(
                     node,
                     right,
@@ -20,7 +20,7 @@ pub fn parse_trim(tokens: Vec<token::Token>, index: &mut usize) -> node::CalculN
             }
             token::TokenKind::TokenDiv => {
                 *index += 1;
-                let right = parse_factor(tokens.clone(), index);
+                let right = parse_factor(tokens.clone(), index)?;
                 node = resp::handler::make_operator_node(
                     node,
                     right,
@@ -29,7 +29,7 @@ pub fn parse_trim(tokens: Vec<token::Token>, index: &mut usize) -> node::CalculN
             }
             token::TokenKind::TokenModulo => {
                 *index += 1;
-                let right = parse_factor(tokens.clone(), index);
+                let right = parse_factor(tokens.clone(), index)?;
                 node = resp::handler::make_operator_node(
                     node,
                     right,
@@ -43,5 +43,5 @@ pub fn parse_trim(tokens: Vec<token::Token>, index: &mut usize) -> node::CalculN
         }
     }
 
-    return node;
+    return Ok(node);
 }